+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to `svelte-adapter-uws` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.6.0-next.94] - 2026-08-25
+## [0.6.0-next.94] - 2026-08-26
 
 <!-- consumer-release-summary:start -->
 ### Consumer summary